前端和后端对接，后端JSONARRAY转obj的时间类型问题，后端id问题，后端socketSession问题

+++ src/cn/jxufe/dao/LandDAO.java
@@ -3,6 +3,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import cn.jxufe.entity.Land;
 public interface LandDAO extends PagingAndSortingRepository<Land, Long> {
-	public Land findByLandId(int landId);
-	public Land findByLandIdAndUId(int landId,long uId);
+	public Land findByLandId(long landId);
+	public Land findByLandIdAndUId(long landId,long uId);
 }
